💥 Rename effects module
Rename effects module to effect to be more consistent with the filter module api semantic
