+++ src/torchfx/__init__.py
@@ -1,6 +1,6 @@
 import torchfx.filter as filter  # noqa: A001,A004
 import torchfx.typing as typing
-from torchfx.effects import FX
+from torchfx.effect import FX
 from torchfx.wave import Wave
 
 __all__ = ["FX", "Wave", "filter", "typing"]
